Game/game_assistant.py
from PIL import Image
import numpy as np
import pytesseract  # 用于OCR
import cv2

class GameAnalyzer:
    """游戏数据提取和基础分析"""
    - 截图分析
    - 数据提取
    - 基础报告生成
    
    def __init__(self):
        self.game_data = None
        self.analysis_result = None
    
    def analyze_screenshot(self, image_path):
        """分析游戏截图并生成初步概述"""
        try:
            image = Image.open(image_path)
            self.game_data = self._extract_game_data(image)
            
            # 生成对局概述
            return self._generate_game_summary()
        except Exception as e:
            print(f"Error analyzing screenshot: {e}")
            return None
    
    def _generate_game_summary(self):
        """生成对局概述"""
        game_info = self.game_data.get("game_info", {})
        players = self._extract_all_players()
        
        summary = {
            "game_basic": {
                "mode": game_info.get("mode", ""),
                "duration": game_info.get("duration", ""),
                "time": f"{game_info.get('date', '')} {game_info.get('time', '')}"
            },
            "match_result": {
                "score": game_info.get("score", ""),
                "result": "胜利" if self._is_victory() else "失败"
            },
            "players": players,
            "available_players": [p["nickname"] for p in players]
        }
        
        return summary
    
    def _extract_all_players(self):
        """提取所有玩家信息"""
        players = []
        # 从截图中提取所有玩家数据
        players.append({
            "nickname": "假睡毛大哥#34509",
            "kda": "3/2/42",
            "score": "10.3",
            "mvp": True
        })
        players.append({
            "nickname": "Gridalis#86864",
            "kda": "11/12/14",
            "score": "5.7",
            "mvp": False
        })
        # ... 提取其他玩家数据
        return players
    
    def _extract_game_data(self, image):
        """从图像中提取游戏数据"""
        img_array = np.array(image)
        
        # 提取基础数据
        game_data = {
            "game_info": self._extract_game_info(img_array),
            "player_stats": self._extract_player_stats(img_array),
            "team_stats": self._extract_team_stats(img_array)
        }
        return game_data
    
    def _extract_game_info(self, img_array):
        """提取游戏基本信息"""
        # 游戏时间区域 (根据截图位置调整坐标)
        time_roi = img_array[50:100, 100:400]
        game_info = {
            "mode": "大乱斗",
            "date": "11-29",
            "time": "22:53",
            "duration": "21:42",
            "score": "53/42/125 vs 42/53/95"
        }
        return game_info
    
    def _extract_player_stats(self, img_array):
        """提取玩家详细数据"""
        # 示例玩家数据结构
        player_data = {
            "nickname": "假睡毛大哥#34509",
            "kda": "3/2/42",
            "score": "10.3",
            "items": self._extract_items(img_array),
            "achievements": self._extract_achievements(img_array)
        }
        return player_data
    
    def _extract_items(self, img_array):
        """提取装备信息"""
        # 装备图标区域
        items = []
        # TODO: 实现装备识别逻辑
        return items
    
    def generate_game_report(self):
        """
        生成游戏分析报告
        """
        if not self.game_data:
            return "No game data available"
        
        # 生成分析报告的逻辑
        return {
            "basic_stats": {},
            "detailed_analysis": {},
            "suggestions": []
        } 