Game/requirements.txt
transformers
torch
datasets
pandas
numpy
python-dotenv
scikit-learn
fastapi==0.68.1
python-multipart==0.0.5
pillow==8.3.2
opencv-python==4.5.3.56
pytesseract==0.3.8
aiofiles==0.7.0
uvicorn==0.15.0 